# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_app_api_handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  backend_route-fix-backup-20250603_161334_handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
...
  request.CreateUserRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      full_name:
        type: string
      is_active:
        description: Pointer to allow explicit true/false, or omit for default
        type: boolean
      is_admin:
        description: 'Added: Admin flag'
        type: boolean
      is_superuser:
        description: Pointer for explicit setting
        type: boolean
      last_name:
        maxLength: 100
        minLength: 1
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
      role_ids:
        description: 'Optional: IDs of roles to assign'
        items:
          type: integer
        type: array
      tenant_id:
        description: 'Added: Tenant association'
        type: integer
      username:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - username
    type: object
...
  request.UpdateUserRequest:
    properties:
      email:
        description: Pointer to allow partial updates
        type: string
      first_name:
        maxLength: 100
        minLength: 1
        type: string
      full_name:
        type: string
      is_active:
        type: boolean
      is_admin:
        description: 'Added: Admin flag'
        type: boolean
      is_superuser:
        type: boolean
      last_name:
        maxLength: 100
        minLength: 1
        type: string
      password:
        description: 'Optional: for password change'
        maxLength: 100
        minLength: 8
        type: string
      role_ids:
        description: For changing role assignments
        items:
          type: integer
        type: array
      username:
        maxLength: 100
        minLength: 3
        type: string
    type: object
...
  response.RoleBase:
    properties:
      code:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  response.UserListResponse:
    properties:
      users:
        items:
          $ref: '#/definitions/response.UserResponse'
        type: array
    type: object
  response.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_admin:
        description: Add this field
        type: boolean
      is_superuser:
        type: boolean
      last_login:
        type: string
      last_name:
        type: string
      roles:
        items:
          $ref: '#/definitions/response.RoleBase'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: reksolindo.opuschamber.com
info:
  contact: {}
  description: Multi-tenant enterprise asset management system
  title: Reksolindo Enterprise Asset Management API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieves a paginated list of users, optionally filtered.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: limit
        type: integer
      - description: Search term for username, email, or full name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            $ref: '#/definitions/response.UserListResponse'
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Allows an administrator to create a new user account.
      parameters:
      - description: User creation details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden - Insufficient permissions
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "409":
          description: Username or email already exists
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user (Admin)
      tags:
      - users
  /users/{id}:
    delete:
      description: Deletes a specific user by their ID (soft delete typically).
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid User ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user by ID
      tags:
      - users
    get:
      description: Retrieves details of a specific user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Invalid User ID
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates details of a specific user by their ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/response.UserResponse'
        "400":
          description: Invalid User ID or request payload
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "409":
          description: Email already in use by another user
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_route-fix-backup-20250603_161334_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user by ID
      tags:
      - users
...
  /users/profile:
    get:
      description: Retrieves the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Current user profile
          schema:
            $ref: '#/definitions/response.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_app_api_handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/backend_app_api_handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/backend_app_api_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
schemes:
- https
- http
swagger: "2.0"